+++ buildacomputer/Models/hard_drives.cs
@@ -7,19 +7,19 @@ namespace buildacomputer.Models
 {
     public class hard_drives
     {
-        public int hard_drive_id { get; set; }
+        public long hard_drive_id { get; set; }
 
         public string hard_drive_name { get; set; }
 
-        public int buffer_kb { get; set; }
+        public long buffer_kb { get; set; }
 
-        public int capacity_gb { get; set; }
+        public long capacity_gb { get; set; }
 
-        public int manufacturer_id { get; set; }
+        public long manufacturer_id { get; set; }
 
-        public int bus_interface_id { get; set; }
+        public long bus_interface_id { get; set; }
 
-        public int drive_bay_width_id { get; set; }
+        public long drive_bay_width_id { get; set; }
 
         public virtual ICollection<l_hard_drives_hard_drive_features> l_hard_drives_hard_drive_features { get; set; }
         public virtual manufacturers manufacturers { get; set; }
