+++ buildacomputer/Models/l_motherboard_bus_interfaces.cs
@@ -7,11 +7,11 @@ namespace buildacomputer.Models
 {
     public class l_motherboard_bus_interfaces
     {
-        public int motherboard_id { get; set; }
+        public long motherboard_id { get; set; }
 
-        public int bus_interface_id { get; set; }
+        public long bus_interface_id { get; set; }
 
-        public int bus_inteface_count { get; set; }
+        public long bus_inteface_count { get; set; }
 
         public virtual motherboards motherboards { get; set; }
         public virtual bus_interfaces bus_interfaces { get; set; }
