+++ buildacomputer/Models/l_motherboards_expansion_slots.cs
@@ -7,11 +7,11 @@ namespace buildacomputer.Models
 {
     public class l_motherboards_expansion_slots
     {
-        public int motherboard_id { get; set; }
+        public long motherboard_id { get; set; }
 
-        public int expansion_slot_id { get; set; }
+        public long expansion_slot_id { get; set; }
 
-        public int expansion_slot_count { get; set; }
+        public long expansion_slot_count { get; set; }
 
         public virtual motherboards motherboards { get; set; }
         public virtual expansion_slots expansion_slots { get; set; }
