+++ buildacomputer/Models/memories.cs
@@ -7,17 +7,17 @@ namespace buildacomputer.Models
 {
     public class memories
     {
-        public int memory_id { get; set; }
+        public long memory_id { get; set; }
 
         public string memory_name { get; set; }
 
-        public int memory_speed_mhz { get; set; }
+        public long memory_speed_mhz { get; set; }
 
-        public int memory_size_mb { get; set; }
+        public long memory_size_mb { get; set; }
 
-        public int manufacturer_id { get; set; }
+        public long manufacturer_id { get; set; }
 
-        public int memory_type_id { get; set; }
+        public long memory_type_id { get; set; }
 
         public virtual memory_types memory_types { get; set; }
         public virtual manufacturers manufacturers { get; set; }
