+++ buildacomputer/Models/motherboard_nb_chipsets.cs
@@ -7,11 +7,11 @@ namespace buildacomputer.Models
 {
     public class motherboard_nb_chipsets
     {
-        public int motherboard_nb_chipset_id { get; set; }
+        public long motherboard_nb_chipset_id { get; set; }
 
         public string motherboard_nb_chipset_name { get; set; }
 
-        public int manufacturer_id { get; set; }
+        public long manufacturer_id { get; set; }
 
         public virtual ICollection<motherboards> motherboards { get; set; }
         public virtual manufacturers manufacturers { get; set; }
