+++ buildacomputer/Models/processor_cores.cs
@@ -7,19 +7,19 @@ namespace buildacomputer.Models
 {
     public class processor_cores
     {
-        public int processor_core_id { get; set; }
+        public long processor_core_id { get; set; }
 
         public string processor_core_name { get; set; }
 
-        public int processor_core_count { get; set; }
+        public long processor_core_count { get; set; }
 
-        public int processor_core_l1cache_kb { get; set; }
+        public long processor_core_l1cache_kb { get; set; }
 
-        public int processor_core_l2cache_kb { get; set; }
+        public long processor_core_l2cache_kb { get; set; }
 
-        public int processor_core_l3cache_kb { get; set; }
+        public long processor_core_l3cache_kb { get; set; }
 
-        public int processor_core_manufacturing_technology_nm { get; set; }
+        public long processor_core_manufacturing_technology_nm { get; set; }
 
         public ICollection<processors> processors { get; set; }
     }
