+++ buildacomputer/Models/sound_cards.cs
@@ -7,15 +7,15 @@ namespace buildacomputer.Models
 {
     public class sound_cards
     {
-        public int sound_card_id { get; set; }
+        public long sound_card_id { get; set; }
 
         public string sound_card_name { get; set; }
 
-        public int manufacturer_id { get; set; }
+        public long manufacturer_id { get; set; }
 
-        public int expansion_slot_id { get; set; }
+        public long expansion_slot_id { get; set; }
 
-        public int sound_chip_id { get; set; }
+        public long sound_chip_id { get; set; }
 
         public virtual manufacturers manufacturers { get; set; }
         public virtual expansion_slots expansion_slots { get; set; }
