+++ buildacomputer/Models/sound_channel_standards.cs
@@ -8,7 +8,7 @@ namespace buildacomputer.Models
 {
     public class sound_channel_standards
     {
-        public int sound_channel_standard_id { get; set; }
+        public long sound_channel_standard_id { get; set; }
 
         public string sound_channel_standard_name { get; set; }
